orchestrator/requirements.txt
typer==0.12.3
GitPython==3.1.43
jinja2==3.1.4
colorama==0.4.6
pyyaml==6.0.2
